feat: add Kadal-typed enemy
Kadal-typed enemy can walk and also attack the player within the range of 4 vlocks (32x4). This type of enemy can also idle, which means the enemy will stop for a certain amount of time when it meets a wall before turning back. Yet, the attack action have not been implemented.